# api-docs.json
{
    "openapi":  "3.0.1",
    "info":  {
                 "title":  "Katron AI Gift Card API Documentation",
                 "version":  "1.0"
             },
    "servers":  [
                    {
                        "url":  "https://api.ktngiftcard.katronai.com/katron-gift-card",
                        "description":  "Generated server url"
                    }
                ],
    "security":  [
                     {
                         "BearerAuth":  [

                                        ]
                     }
                 ],
    "paths":  {
...
                  "/api/user/addCardToPayArc":  {
                                                    "post":  {
                                                                 "tags":  [
                                                                              "user-controller"
                                                                          ],
                                                                 "operationId":  "addCardToPayArc",
                                                                 "requestBody":  {
                                                                                     "content":  {
                                                                                                     "application/json":  {
                                                                                                                              "schema":  {
                                                                                                                                             "$ref":  "#/components/schemas/CardCreateRequestModel"
                                                                                                                                         }
                                                                                                                          }
                                                                                                 },
                                                                                     "required":  true
                                                                                 },
                                                                 "responses":  {
                                                                                   "200":  {
                                                                                               "description":  "OK",
                                                                                               "content":  {
                                                                                                               "*/*":  {
                                                                                                                           "schema":  {
                                                                                                                                          "$ref":  "#/components/schemas/Response"
                                                                                                                                      }
                                                                                                                       }
                                                                                                           }
                                                                                           }
                                                                               }
                                                             }
                                                },
...
                  "/api/giftCards/createGiftCardOrder":  {
                                                             "post":  {
                                                                          "tags":  [
                                                                                       "gift-card-controller"
                                                                                   ],
                                                                          "operationId":  "createGiftCardOrder",
                                                                          "requestBody":  {
                                                                                              "content":  {
                                                                                                              "application/json":  {
                                                                                                                                       "schema":  {
                                                                                                                                                      "$ref":  "#/components/schemas/GiftcardOrderRequest"
                                                                                                                                                  }
                                                                                                                                   }
                                                                                                          },
                                                                                              "required":  true
                                                                                          },
                                                                          "responses":  {
                                                                                            "200":  {
                                                                                                        "description":  "OK",
                                                                                                        "content":  {
                                                                                                                        "*/*":  {
                                                                                                                                    "schema":  {
                                                                                                                                                   "$ref":  "#/components/schemas/Response"
                                                                                                                                               }
                                                                                                                                }
                                                                                                                    }
                                                                                                    }
                                                                                        }
                                                                      }
                                                         },
...
                  "/api/auth/verifyEmailOtp":  {
                                                   "post":  {
                                                                "tags":  [
                                                                             "auth-controller"
                                                                         ],
                                                                "operationId":  "verifyEmailOtp",
                                                                "parameters":  [
                                                                                   {
                                                                                       "name":  "otp",
                                                                                       "in":  "query",
                                                                                       "required":  true,
                                                                                       "schema":  {
                                                                                                      "type":  "string"
                                                                                                  }
                                                                                   }
                                                                               ],
                                                                "responses":  {
                                                                                  "200":  {
                                                                                              "description":  "OK",
                                                                                              "content":  {
                                                                                                              "*/*":  {
                                                                                                                          "schema":  {
                                                                                                                                         "$ref":  "#/components/schemas/Response"
                                                                                                                                     }
                                                                                                                      }
                                                                                                          }
                                                                                          }
                                                                              }
                                                            }
                                               },
...
                  "/api/auth/login":  {
                                          "post":  {
                                                       "tags":  [
                                                                    "auth-controller"
                                                                ],
                                                       "operationId":  "login",
                                                       "parameters":  [
                                                                          {
                                                                              "name":  "email",
                                                                              "in":  "query",
                                                                              "required":  true,
                                                                              "schema":  {
                                                                                             "type":  "string"
                                                                                         }
                                                                          },
                                                                          {
                                                                              "name":  "password",
                                                                              "in":  "query",
                                                                              "required":  true,
                                                                              "schema":  {
                                                                                             "type":  "string"
                                                                                         }
                                                                          }
                                                                      ],
                                                       "responses":  {
                                                                         "200":  {
                                                                                     "description":  "OK",
                                                                                     "content":  {
                                                                                                     "*/*":  {
                                                                                                                 "schema":  {
                                                                                                                                "$ref":  "#/components/schemas/Response"
                                                                                                                            }
                                                                                                             }
                                                                                                 }
                                                                                 }
                                                                     }
                                                   }
                                      },
                  "/api/auth/forgotPassword":  {
                                                   "post":  {
                                                                "tags":  [
                                                                             "auth-controller"
                                                                         ],
                                                                "operationId":  "forgotPassword",
                                                                "parameters":  [
                                                                                   {
                                                                                       "name":  "email",
                                                                                       "in":  "query",
                                                                                       "required":  true,
                                                                                       "schema":  {
                                                                                                      "type":  "string"
                                                                                                  }
                                                                                   }
                                                                               ],
                                                                "responses":  {
                                                                                  "200":  {
                                                                                              "description":  "OK",
                                                                                              "content":  {
                                                                                                              "*/*":  {
                                                                                                                          "schema":  {
                                                                                                                                         "$ref":  "#/components/schemas/Response"
                                                                                                                                     }
                                                                                                                      }
                                                                                                          }
                                                                                          }
                                                                              }
                                                            }
                                               },
...
              },
    "components":  {
                       "schemas":  {
                                       "Response":  {
                                                        "type":  "object",
                                                        "properties":  {
                                                                           "status":  {
                                                                                          "type":  "integer",
                                                                                          "format":  "int32"
                                                                                      },
                                                                           "message":  {
                                                                                           "type":  "string"
                                                                                       },
                                                                           "data":  {
                                                                                        "type":  "object"
                                                                                    }
                                                                       }
                                                    },
...
                                       "CardCreateRequestModel":  {
                                                                      "required":  [
                                                                                       "cardExpiryDate",
                                                                                       "cardIssuerType",
                                                                                       "cardNumber",
                                                                                       "cardholderName",
                                                                                       "cvv"
                                                                                   ],
                                                                      "type":  "object",
                                                                      "properties":  {
                                                                                         "cardNumber":  {
                                                                                                            "type":  "string"
                                                                                                        },
                                                                                         "cardExpiryDate":  {
                                                                                                                "type":  "string"
                                                                                                            },
                                                                                         "cardholderName":  {
                                                                                                                "type":  "string"
                                                                                                            },
                                                                                         "cardIssuerType":  {
                                                                                                                "type":  "string"
                                                                                                            },
                                                                                         "cvv":  {
                                                                                                     "type":  "string"
                                                                                                 }
                                                                                     }
                                                                  },
...
                                       "GiftcardOrderRequest":  {
                                                                    "required":  [
                                                                                     "failureUrl",
                                                                                     "successUrl"
                                                                                 ],
                                                                    "type":  "object",
                                                                    "properties":  {
                                                                                       "giftCardId":  {
                                                                                                          "type":  "integer",
                                                                                                          "format":  "int64"
                                                                                                      },
                                                                                       "quantity":  {
                                                                                                        "type":  "integer",
                                                                                                        "format":  "int32"
                                                                                                    },
                                                                                       "unitPrice":  {
                                                                                                         "type":  "number",
                                                                                                         "format":  "float"
                                                                                                     },
                                                                                       "successUrl":  {
                                                                                                          "type":  "string"
                                                                                                      },
                                                                                       "failureUrl":  {
                                                                                                          "type":  "string"
                                                                                                      }
                                                                                   }
                                                                },
...
                                   },
                       "securitySchemes":  {
                                               "BearerAuth":  {
                                                                  "type":  "http",
                                                                  "scheme":  "bearer",
                                                                  "bearerFormat":  "JWT"
                                                              }
                                           }
                   }
}